requirements.txt
Flask
PyPDF2
firebase-admin
pinecone-client
langchain